Increase the dtype capacity of occurrence counts
Switch from uint16 to uint32 to allow counts larger than 2^16.

+++ weight/train.py
@@ -31,7 +31,7 @@ from weight import vampire
 log = logging.getLogger(__name__)
 
 
-def occurrence_count_vector(token_counts, symbol_name_to_index, dtype=np.uint16):
+def occurrence_count_vector(token_counts, symbol_name_to_index, dtype=np.uint32):
     """
     0. variable
     1. negation
